backend/requirements/base.txt
# backend/requirements/base.txt
# Base Dependencies for Local & Production Environments

# Core Framework
Django==5.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1

# Async / WebSocket
channels==4.0.0
channels-redis==4.1.0

# Logging
structlog==22.1.0

# Background Tasks & Scheduling
celery==5.3.4
django-celery-beat>=2.6.1
redis==5.0.1
django-redis==5.4.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
websockets==12.0
numpy>=1.26
pandas>=2.2
ccxt==4.1.2
cryptography==41.0.7

# Debug Tools (safe for local)
django-debug-toolbar==4.4.6
